Move create_mock_tests to conftest and add test for test_trade-history
